Make the 's' key print the filename
As opposed to the 's' key completely duplicating the current image file,
make pressing the 's' merely print the file name to standard output.

Then, the filenames can be redirected to a text file,
which can be used as a photo album (with feh), or with shell
scripts to duplicate the image files.